Model-checked abstract_factory implementation.
The main change to this implementation is that we store a pointer to the
implementation details in the abstract_factory instead of a copy of the
structure. This is due to some limitations in CBMC.

Furthermore, we can't model check dynamic growth of the abstract factory.  We
have unit testing in place for this, but this is a current limitation of CBMC.